+++ lua/render-markdown/config/anti_conceal.lua
@@ -1,8 +1,9 @@
 ---@class (exact) render.md.anti.conceal.Config
 ---@field enabled boolean
----@field ignore render.md.conceal.Ignore
+---@field disabled_modes render.md.Modes
 ---@field above integer
 ---@field below integer
+---@field ignore render.md.conceal.Ignore
 
 ---@alias render.md.conceal.Ignore table<render.md.Element, render.md.Modes>
 
@@ -32,6 +33,12 @@ local M = {}
 M.default = {
     -- This enables hiding any added text on the line the cursor is on.
     enabled = true,
+    -- Modes to disable anti conceal feature.
+    disabled_modes = false,
+    -- Number of lines above cursor to show.
+    above = 0,
+    -- Number of lines below cursor to show.
+    below = 0,
     -- Which elements to always show, ignoring anti conceal behavior. Values can either be
     -- booleans to fix the behavior or string lists representing modes where anti conceal
     -- behavior will be ignored. Valid values are:
@@ -41,23 +48,20 @@ M.default = {
         code_background = true,
         sign = true,
     },
-    -- Number of lines above cursor to show.
-    above = 0,
-    -- Number of lines below cursor to show.
-    below = 0,
 }
 
 ---@param spec render.md.debug.ValidatorSpec
 function M.validate(spec)
     spec:type('enabled', 'boolean')
+    spec:list('disabled_modes', 'string', 'boolean')
+    spec:type('above', 'number')
+    spec:type('below', 'number')
     spec:nested('ignore', function(ignore)
         for _, element in pairs(Element) do
             ignore:list(element, 'string', { 'boolean', 'nil' })
         end
         ignore:check()
     end)
-    spec:type('above', 'number')
-    spec:type('below', 'number')
     spec:check()
 end