+++ lua/render-markdown/config/checkbox.lua
@@ -63,8 +63,7 @@ M.default = {
 
 ---@param spec render.md.debug.ValidatorSpec
 function M.validate(spec)
-    local Base = require('render-markdown.config.base')
-    Base.validate(spec)
+    require('render-markdown.config.base').validate(spec)
     spec:type('bullet', 'boolean')
     spec:type('right_pad', 'number')
     spec:nested({ 'unchecked', 'checked' }, function(box)