+++ lua/render-markdown/config/dash.lua
@@ -32,8 +32,7 @@ M.default = {
 
 ---@param spec render.md.debug.ValidatorSpec
 function M.validate(spec)
-    local Base = require('render-markdown.config.base')
-    Base.validate(spec)
+    require('render-markdown.config.base').validate(spec)
     spec:type('icon', 'string')
     spec:one_of('width', { 'full' }, 'number')
     spec:type('left_margin', 'number')