+++ lua/render-markdown/config/html.lua
@@ -37,8 +37,7 @@ M.default = {
 
 ---@param spec render.md.debug.ValidatorSpec
 function M.validate(spec)
-    local Base = require('render-markdown.config.base')
-    Base.validate(spec)
+    require('render-markdown.config.base').validate(spec)
     spec:nested('comment', function(comment)
         comment:type('conceal', 'boolean')
         comment:type('text', { 'string', 'nil' })