+++ lua/render-markdown/config/indent.lua
@@ -32,8 +32,7 @@ M.default = {
 
 ---@param spec render.md.debug.ValidatorSpec
 function M.validate(spec)
-    local Base = require('render-markdown.config.base')
-    Base.validate(spec)
+    require('render-markdown.config.base').validate(spec)
     spec:type('per_level', 'number')
     spec:type('skip_level', 'number')
     spec:type('skip_heading', 'boolean')