+++ lua/render-markdown/config/inline_highlight.lua
@@ -19,8 +19,7 @@ M.default = {
 
 ---@param spec render.md.debug.ValidatorSpec
 function M.validate(spec)
-    local Base = require('render-markdown.config.base')
-    Base.validate(spec)
+    require('render-markdown.config.base').validate(spec)
     spec:type('highlight', 'string')
     spec:check()
 end