+++ lua/render-markdown/config/latex.lua
@@ -36,8 +36,7 @@ M.default = {
 
 ---@param spec render.md.debug.ValidatorSpec
 function M.validate(spec)
-    local Base = require('render-markdown.config.base')
-    Base.validate(spec)
+    require('render-markdown.config.base').validate(spec)
     spec:type('converter', 'string')
     spec:type('highlight', 'string')
     spec:one_of('position', vim.tbl_values(Position))