+++ lua/render-markdown/config/link.lua
@@ -102,8 +102,7 @@ M.default = {
 
 ---@param spec render.md.debug.ValidatorSpec
 function M.validate(spec)
-    local Base = require('render-markdown.config.base')
-    Base.validate(spec)
+    require('render-markdown.config.base').validate(spec)
     spec:nested('footnote', function(footnote)
         footnote:type('enabled', 'boolean')
         footnote:type('superscript', 'boolean')