+++ lua/render-markdown/lib/config.lua
@@ -81,8 +81,7 @@ end
 -- stylua: ignore
 ---@param spec render.md.debug.ValidatorSpec
 function Config.validate(spec)
-    local Base = require('render-markdown.config.base')
-    Base.validate(spec)
+    require('render-markdown.config.base').validate(spec)
     spec:type('max_file_size', 'number')
     spec:type('debounce', 'number')
     spec:nested('anti_conceal', require('render-markdown.config.anti_conceal').validate)