+++ lua/render-markdown/render/markdown/heading.lua
@@ -12,7 +12,7 @@ local colors = require('render-markdown.core.colors')
 ---@field sign? string
 ---@field fg? string
 ---@field bg? string
----@field width render.md.base.Width
+---@field width render.md.heading.Width
 ---@field left_margin number
 ---@field left_pad number
 ---@field right_pad number