+++ lua/render-markdown/types.lua
@@ -89,7 +89,7 @@
 ---@field language_icon? boolean
 ---@field language_name? boolean
 ---@field disable_background? boolean|string[]
----@field width? render.md.base.Width
+---@field width? render.md.code.Width
 ---@field left_margin? number
 ---@field left_pad? number
 ---@field right_pad? number
@@ -139,7 +139,7 @@
 ---@field icons? render.md.heading.String
 ---@field position? render.md.heading.Position
 ---@field signs? string[]
----@field width? render.md.base.Width|(render.md.base.Width)[]
+---@field width? render.md.heading.Width|(render.md.heading.Width)[]
 ---@field left_margin? number|number[]
 ---@field left_pad? number|number[]
 ---@field right_pad? number|number[]