+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   [a1b0988](https://github.com/MeanderingProgrammer/render-markdown.nvim/commit/a1b0988f5ab26698afb56b9c2f0525a4de1195c1)
 - include highlight query files in checkhealth [4f05de7](https://github.com/MeanderingProgrammer/render-markdown.nvim/commit/4f05de7536571e231b66dd0363af347b162b5fd7)
 - pad suffix of latex so backgrounds are aligned [a212596](https://github.com/MeanderingProgrammer/render-markdown.nvim/commit/a2125961287c78dfc03cb628b7bed7f91b70fd45)
+- allow anti conceal to be disabled in specific modes [dd62056](https://github.com/MeanderingProgrammer/render-markdown.nvim/commit/dd620561843d5a62c917fffe77f2923c965d12c7)
 
 ### Bug Fixes